Things should be working now

+++ parse_wvb.py
@@ -3,8 +3,14 @@ import os
 import struct
 import sys
 import subprocess
+import queue
+import threading
 
 import pydub
+import common
+
+def convert_vgmstream(input_filename, output_filename):
+    subprocess.call('vgmstream_cli.exe -q -o "%s" "%s"' % (output_filename, input_filename))
 
 
 def percentage_to_db(percentage):
@@ -17,12 +23,14 @@ def percentage_to_db(percentage):
 
 def find_num_samples(infile, offset):
     cur_offset = infile.tell()
-    infile.seek(offset)
+    infile.seek(0, 2)
+    filelen = infile.tell()
+    infile.seek(offset + 16)
 
     # TODO: This could be a potential source of problems
     # vgmstream also uses these signatures to detect the end frame, so it should be ok I think?
     filesize = 0
-    while infile.read(16) not in [b'\x00\x07' + b'\x77' * 14, b'\x00\x07' + b'\x00' * 14]:
+    while infile.tell() < filelen and infile.read(16) not in [b'\x00\x07' + b'\x77' * 14, b'\x00\x07' + b'\x00' * 14, b'\00' * 16]:
         filesize += 16
 
     infile.seek(cur_offset)
@@ -30,64 +38,85 @@ def find_num_samples(infile, offset):
     return filesize
 
 
-def parse_wvb_old(infile, output_folder):
+def parse_wvb_old(infile, output_folder, num_threads):
     archive_size, section_size, section2_relative_offset = struct.unpack("<III", infile.read(12))
     section2_offset = archive_size - section2_relative_offset
 
     infile.seek(0x20)
 
     file_id = 0
+    files = []
 
     while infile.tell() < 0x4000:
         entry_id, _, _, volume, pan, entry_type, sample_rate1, sample_rate2, offset1, offset2, filesize, _ = struct.unpack("<HHBBBBIIIIII", infile.read(0x20))
 
         cur_offset = infile.tell()
 
-        files = []
+        entry_type &= 0x0f
 
         if entry_type == 0:
             continue
 
         elif entry_type == 2:
             offset1 -= 0xf020
-            files.append([offset1, sample_rate1, find_num_samples(infile, offset1)])
+            files.append([entry_id, offset1, sample_rate1, find_num_samples(infile, offset1), volume, pan])
 
         elif entry_type == 4:
-            files.append([offset1 - 0xf020, sample_rate1, offset2 - offset1])
+            files.append([entry_id, offset1 - 0xf020, sample_rate1, offset2 - offset1, volume, pan])
 
         elif entry_type == 3:
-            files.append([section2_offset + offset1, sample_rate1, filesize])
+            files.append([entry_id, section2_offset + offset1, sample_rate1, filesize, volume, pan])
 
         else:
-            print("Unknown entry type:", entry_type, "%02x" % (infile.tell() - 0x20))
+            print("Unknown entry type:", entry_type, "%02x" % (infile.tell() - 0x20), output_folder)
             exit(1)
 
+        infile.seek(cur_offset)
 
-        for offset, sample_rate, size in files:
-            output_filename = os.path.join(output_folder, "%04d.pcm" % entry_id)
+    def thread_worker():
+        while True:
+            item = queue_data.get()
+
+            if item is None:
+                break
+
+            output_filename, offset, sample_rate, size, volume, pan = item
 
             print("Extracting", output_filename)
 
-            with open(output_filename, "wb") as outfile:
-                outfile.write(struct.pack(">IHHB", size, 0, sample_rate, 1))
-                outfile.write(bytearray([0] * 0x7f7))
-
-                infile.seek(offset)
-                outfile.write(infile.read(size))
-
             wav_output_filename = output_filename.replace(".pcm", ".wav")
-            subprocess.call('vgmstream_cli.exe -o "%s" "%s"' % (wav_output_filename, output_filename))
-            os.unlink(output_filename)
+            convert_vgmstream(output_filename, wav_output_filename)
 
             wav = pydub.AudioSegment.from_file(wav_output_filename)
             wav = wav.pan((pan - (128 / 2)) / (128 / 2))
             wav += percentage_to_db((volume / 127) * 100 * 0.75)
             wav.export(wav_output_filename, format="wav")
 
-        infile.seek(cur_offset)
+            queue_data.task_done()
+
+    queue_data = queue.Queue()
+    for entry_id, offset, sample_rate, size, volume, pan in files:
+        output_filename = os.path.join(output_folder, "%04d.pcm" % entry_id)
+
+        with open(output_filename, "wb") as outfile:
+            outfile.write(struct.pack(">IHHB", size, 0, sample_rate, 1))
+            outfile.write(bytearray([0] * 0x7f7))
+
+            infile.seek(offset)
+            outfile.write(infile.read(size))
+
+        queue_data.put((output_filename, offset, sample_rate, size, volume, pan))
+
+    common.process_queue(queue_data, thread_worker, num_threads)
+
+    for entry_id, _, _, _, _, _ in files:
+        output_filename = os.path.join(output_folder, "%04d.pcm" % entry_id)
+
+        if os.path.exists(output_filename):
+            os.unlink(output_filename)
 
 
-def parse_wvb_new(infile, output_folder):
+def parse_wvb_new(infile, output_folder, num_threads):
     header, keysound_table_offset = struct.unpack("<II", infile.read(8))
     infile.seek(0x10, 0)
 
@@ -119,13 +148,14 @@ def parse_wvb_new(infile, output_folder):
             'offset': offset,
             'filesize': filesize,
             'unk1': unk1, # Does one of these relate to the sample rate somehow?
-            'unk2': unk2,
+            'unk2': unk2 >> 8,
+            'unk2o': unk2,
             'unk3': unk3,
         })
 
     data_offset = infile.tell()
 
-
+    files = []
     for entry in keysound_entries:
         entry_id = entry['entry_id']
         size = file_entries[entry['file_id']]['filesize']
@@ -135,19 +165,67 @@ def parse_wvb_new(infile, output_folder):
         sample_rate = 44100
         channels = 1
 
-        if file_entries[entry['file_id']]['unk2'] & 0xff == 0x44:
-            sample_rate = 44100
+        # Thanks Konami
+        sample_rates = {
+            0x3e: 44100,
+            0x3f: 42000,
+            0x40: 40000,
+            0x41: 38000,
+            0x42: 36000,
+            0x43: 34000,
+            0x44: 32000,
+            0x45: 30000,
+            0x46: 28000,
+            0x47: 26000,
+            0x48: 24000,
+            0x4a: 22050,
+            0x4c: 20000,
+            0x4d: 18000,
+            0x50: 16000,
+            # 0x51: 15000,
+            # 0x52: 14000,
+            # 0x53: 13000,
+            0x54: 12000,
+        }
 
-        elif file_entries[entry['file_id']]['unk2'] & 0xff == 0x7d:
-            sample_rate = 32000
+        if file_entries[entry['file_id']]['unk2'] in sample_rates:
+            sample_rate = sample_rates[file_entries[entry['file_id']]['unk2']]
 
         else:
-            print("Unknown sample rate", file_entries[entry['file_id']]['unk2'])
+            print("Unknown sample rate %04x" % file_entries[entry['file_id']]['unk2o'], output_folder)
             exit(1)
 
-        output_filename = os.path.join(output_folder, "%04d.pcm" % entry_id)
+        files.append((entry_id, offset, sample_rate, size, volume, pan))
 
-        print("Extracting", output_filename)
+    def thread_worker():
+        while True:
+            item = queue_data.get()
+
+            if item is None:
+                break
+
+            output_filename, offset, sample_rate, size, volume, pan = item
+
+            print("Extracting", output_filename)
+
+            wav_output_filename = output_filename.replace(".pcm", ".wav")
+            convert_vgmstream(output_filename, wav_output_filename)
+            wav = pydub.AudioSegment.from_file(wav_output_filename)
+
+            if entry['mix'] == 2:
+                wav = wav.set_channels(2).pan(((pan - (128 / 2)) / (128 / 2)))
+                wav += percentage_to_db((volume / 100) * 100 * 0.75)
+
+            else:
+                wav += percentage_to_db((volume / 127) * 100 * 0.75)
+
+            wav.export(wav_output_filename, format="wav")
+
+            queue_data.task_done()
+
+    queue_data = queue.Queue()
+    for entry_id, offset, sample_rate, size, volume, pan in files:
+        output_filename = os.path.join(output_folder, "%04d.pcm" % entry_id)
 
         with open(output_filename, "wb") as outfile:
             outfile.write(struct.pack("<IIII", 0x08640001, 0, 0x800, size))
@@ -158,25 +236,22 @@ def parse_wvb_new(infile, output_folder):
             infile.seek(offset)
             outfile.write(infile.read(size))
 
-        wav_output_filename = output_filename.replace(".pcm", ".wav")
-        subprocess.call('vgmstream_cli.exe -o "%s" "%s"' % (wav_output_filename, output_filename))
-        os.unlink(output_filename)
+        queue_data.put((output_filename, offset, sample_rate, size, volume, pan))
 
-        wav = pydub.AudioSegment.from_file(wav_output_filename)
+    common.process_queue(queue_data, thread_worker, num_threads)
 
-        if entry['mix'] == 2:
-            wav = wav.set_channels(2).pan(((pan - (128 / 2)) / (128 / 2)))
-            wav += percentage_to_db((volume / 100) * 100 * 0.75)
+    for entry_id, _, _, _, _, _ in files:
+        output_filename = os.path.join(output_folder, "%04d.pcm" % entry_id)
 
-        else:
-            wav += percentage_to_db((volume / 127) * 100 * 0.75)
-
-        wav.export(wav_output_filename, format="wav")
+        if os.path.exists(output_filename):
+            os.unlink(output_filename)
 
 
-def parse_wvb(filename, output_folder):
+def parse_wvb(filename, output_folder, threads=4):
     if not os.path.exists(output_folder):
         os.makedirs(output_folder)
+
+    print("Processing", filename)
 
     with open(filename, "rb") as infile:
         header = struct.unpack("<I", infile.read(4))[0]
@@ -186,16 +261,17 @@ def parse_wvb(filename, output_folder):
         infile.seek(0, 0)
 
         if header == filelen:
-            parse_wvb_old(infile, output_folder)
+            parse_wvb_old(infile, output_folder, threads)
 
         else:
-            parse_wvb_new(infile, output_folder)
+            parse_wvb_new(infile, output_folder, threads)
 
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument('--input', help='Input .wvb file', required=True)
     parser.add_argument('--output', help='Output folder', required=True)
+    parser.add_argument('--threads', help='Number of threads to use during extraction', default=4, type=int)
     args = parser.parse_args()
 
-    parse_wvb(args.input, args.output)
+    parse_wvb(args.input, args.output, args.threads)
